+++ Changes.md
@@ -2,6 +2,7 @@ Version Information
 ===================
 Version 405.1.0
   1. Moodle 4.5 version.
+  2. Content to a template.
 
 Version 29.1.8
   1. Update to ES6 module.